+++ TrustedTypes.js
@@ -1,18 +1,99 @@
+import { supported as isSupported } from './trust.js';
+import { events } from './attributes.js';
+import { getMetaCSP, getHTTPCSP } from './http.js';
 /**
  * @See https://github.com/w3c/webappsec-trusted-types/blob/main/src/trustedtypes.js
  * @See https://developer.mozilla.org/en-US/docs/Web/API/Trusted_Types_API
  */
-import { supported as isSupported } from './trust.js';
-import { events } from './attributes.js';
 
-// Maps to `trusted-types 'allow-duplicates'`
-const allowDuplicates = document.documentElement.dataset.allowDuplicates === '1';
+// `<html data-fetch-trust-csp="1">` will enable using actual CSP
+const fetchCSP = document.documentElement.dataset.fetchTrustCsp === '1';
 
-// Maps to headers: `trusted-types <policy-name>`
-const allowedPolicies = new Set([
-	'empty#html', 'empty#script',
-	...document.documentElement.dataset.trustedPolicies.split(' '),
-]);
+function parseCSP({ 'trusted-types': trustedTypes = null } = {}) {
+	if (Array.isArray(trustedTypes)) {
+		const policies = new Set(trustedTypes);
+		const allowDuplicates = policies.has('\'allow-duplicates\'');
+		const hasNone = policies.has('\'none\'');
+
+		if (allowDuplicates) {
+			policies.delete('\'allow-duplicates\'');
+		}
+
+		if (hasNone) {
+			policies.clear();
+		}
+
+		policies.add('empty#html');
+		policies.add('empty#script');
+
+		return { policies, allowDuplicates, hasNone };
+	} else {
+		return {};
+	}
+}
+
+async function getActualCSP() {
+	const { policies, allowDuplicates = false, hasNone = false } = await getHTTPCSP().then(parseCSP);
+
+	if (policies instanceof Set) {
+		return { policies, allowDuplicates, hasNone };
+	} else {
+		return {};
+	}
+}
+
+function init() {
+	// Temporary config if/while fetching actual CSP
+	// Maps to `trusted-types 'allow-duplicates'`
+	const getAllowDuplicates = () => {
+		const csp = parseCSP(getMetaCSP());
+
+		if (typeof csp === 'object' && 'allowDuplicates' in csp) {
+			return csp.allowDuplicates;
+		} else {
+			return document.documentElement.dataset.allowTrustDuplicates === '1';
+		}
+	};
+
+	const getAllowedPolicies = () => {
+		if (document.documentElement.dataset.hasOwnProperty('trustedPolicies')) {
+			return new Set([
+				'empty#html', 'empty#script',
+				...document.documentElement.dataset.trustedPolicies.split(' '),
+			]);
+		} else {
+			const { policies } = parseCSP(getMetaCSP());
+
+			if (policies instanceof Set) {
+				return policies;
+			} else {
+				return new Set(['empty#html', 'empty#script']);
+			}
+		}
+	};
+
+	const fromCSP = parseCSP(getMetaCSP());
+
+	if (typeof fromCSP === 'object' && fromCSP.policies instanceof Set) {
+		const { policies: allowedPolicies, allowDuplicates } = fromCSP;
+		return { allowDuplicates, allowedPolicies };
+	} else {
+		const allowDuplicates = getAllowDuplicates();
+		const allowedPolicies = getAllowedPolicies();
+
+		if (allowedPolicies.has('\'none\'')) {
+			allowedPolicies.clear();
+		}
+
+		allowedPolicies.add('empty#html');
+		allowedPolicies.add('empty#script');
+
+		return { allowDuplicates, allowedPolicies };
+	}
+}
+
+const { allowDuplicates, allowedPolicies } = init();
+
 /**
  * [supported description]
  */
@@ -28,6 +109,8 @@ const symbols = {
 	emptyHTML: Symbol('policy-empty#html'),
 	emptyScript: Symbol('policy-empty#script'),
 	policy: Symbol.for('trust-policy'),
+	defaultPolicy: Symbol('default-policy'),
+	trustedTypesCSP: Symbol('trusted-types-csp'),
 };
 
 if (! Symbol.hasOwnProperty('toStringTag')) {
@@ -41,7 +124,7 @@ function getPolicy(name) {
 }
 
 function hasPolicy(name) {
-	return policies.some(policy => policy.name === name);
+	return getPolicy(name) !== null;
 }
 
 /**
@@ -59,7 +142,7 @@ function getUnsetPolicyException(policy, method) {
  * [TrustedType description]
  * @type {TrustedType}
  */
-class TrustedType {
+export class TrustedType {
 	/**
 	 * [constructor description]
 	 * @param {String} value  [description]
@@ -152,7 +235,7 @@ export class TrustedTypePolicy {
 	 * @param {Function} createScriptURL  [description]
 	 * @param {String} key              [description]
 	 */
-	constructor(name, { createHTML, createScript, createScriptURL }, { key }) {
+	constructor(name, { createHTML, createScript, createScriptURL } = {}, { key }) {
 		// @TODO: Chrome seems to allow any same-origin scripts creating policies
 		if (key !== symbols.trustedKey) {
 			throw new TypeError('Invalid constructor');
@@ -212,6 +295,8 @@ export class BeforeCreatePolicyEvent extends Event {
 
 		if (key !== symbols.trustedKey) {
 			throw new TypeError('Invalid constructor');
+		} else if (! (policy instanceof TrustedTypePolicy)) {
+			throw new TypeError('Not a TrustedTypePolicy');
 		}
 
 		this.policyName = policy.name;
@@ -220,9 +305,9 @@ export class BeforeCreatePolicyEvent extends Event {
 
 /**
  * [enumerable description]
- * @type {TrustedTypeFactory}
+ * @type {TrustedTypePolicyFactory}
  */
-export class TrustedTypeFactory extends EventTarget {
+export class TrustedTypePolicyFactory extends EventTarget {
 	/**
 	 * [constructor description]
 	 * @param {Symbol} key  [description]
@@ -232,6 +317,19 @@ export class TrustedTypeFactory extends EventTarget {
 
 		if (key !== symbols.trustedKey) {
 			throw new TypeError('Invalid constructor');
+		}
+
+		if (fetchCSP) {
+			getActualCSP().then(({ policies, allowDuplicates = false, hasNone = false }) => {
+				if (typeof policies !== 'undefined') {
+					Object.defineProperty(TrustedTypePolicyFactory, symbols.trustedTypesCSP, {
+						value: { policies, allowDuplicates, hasNone },
+						enumberable: false,
+						writable: false,
+						configurable: false,
+					});
+				}
+			}).catch(err => console.error(err));
 		}
 
 		Object.defineProperties(this, {
@@ -253,12 +351,6 @@ export class TrustedTypeFactory extends EventTarget {
 				writable: false,
 				value: this.createPolicy('empty#script', { createScript: () => '' }),
 			},
-		});
-
-		this.addEventListener('beforecreatepolicy', event => {
-			if (this.onbeforecreatepolicy instanceof Function) {
-				this.onbeforecreatepolicy.call(this, event);
-			}
 		});
 	}
 
@@ -296,25 +388,40 @@ export class TrustedTypeFactory extends EventTarget {
 	 * @param  {Function} createScript                  [description]
 	 * @param  {Function} createScriptURL               [description]
 	 */
-	createPolicy(name, { createHTML, createScript, createScriptURL }) {
+	createPolicy(name, { createHTML, createScript, createScriptURL } = {}) {
+		const policy = new TrustedTypePolicy(name, { createHTML, createScript, createScriptURL }, { key: symbols.trustedKey });
+		this.dispatchEvent(new BeforeCreatePolicyEvent('beforecreatepolicy', { policy, key: symbols.trustedKey }));
+
 		if (! name.toString().match(/^[-#a-zA-Z0-9=_/@.%]+$/g)) {
 			throw new TypeError(`Failed to execute 'createPolicy' on 'TrustedTypePolicyFactory': Policy: "${name}" contains invalid characters.`);
-		} else if (allowedPolicies.size > 2 && ! allowedPolicies.has(name)) {
-			throw new TypeError(`Failed to execute 'createPolicy' on 'TrustedTypePolicyFactory': Policy: "${name}" disallowed.`);
-		} else if(! allowDuplicates && hasPolicy(name)) {
-			throw new TypeError(`Failed to execute 'createPolicy' on 'TrustedTypePolicyFactory': Policy: "${name}" already exists.`);
-		} else {
-			const policy = new TrustedTypePolicy(name, { createHTML, createScript, createScriptURL }, { key: symbols.trustedKey });
-			this.dispatchEvent(new BeforeCreatePolicyEvent('beforecreatepolicy', { policy, key: symbols.trustedKey }));
+		} else if (TrustedTypePolicyFactory.hasOwnProperty(symbols.trustedTypesCSP)) {
+			// This requires `fetch()` and parsing response headers, so will not be immediately available
+			const { policies, allowDuplicates, hasNone } = TrustedTypePolicyFactory[symbols.trustedTypesCSP];
 
-			if (policy.name === 'default') {
-				this[symbols.defaultPolicy] = policy;
+			// Will always allow `empty#html` and `empty#script`
+			if (hasNone && name !== 'empty#html' && name !== 'empty#script') {
+				throw new TypeError(`Failed to execute 'createPolicy' on 'TrustedTypePolicyFactory': Policy: "${name}" disallowed.`);
+			} else if (! policies.has(name)) {
+				throw new TypeError(`Failed to execute 'createPolicy' on 'TrustedTypePolicyFactory': Policy: "${name}" disallowed.`);
+			} else if(! allowDuplicates && hasPolicy(name)) {
+				throw new TypeError(`Failed to execute 'createPolicy' on 'TrustedTypePolicyFactory': Policy: "${name}" already exists.`);
 			}
-
-			policies.push(policy);
-
-			return policy;
+		} else {
+			// This is the fallback for before CSP headers are fetched and uses `document.documentElement.dataset`
+			if (allowedPolicies.size > 2 && ! allowedPolicies.has(name)) {
+				throw new TypeError(`Failed to execute 'createPolicy' on 'TrustedTypePolicyFactory': Policy: "${name}" disallowed.`);
+			} else if(! allowDuplicates && hasPolicy(name)) {
+				throw new TypeError(`Failed to execute 'createPolicy' on 'TrustedTypePolicyFactory': Policy: "${name}" already exists.`);
+			}
 		}
+
+		if (policy.name === 'default') {
+			this[symbols.defaultPolicy] = policy;
+		}
+
+		policies.push(policy);
+
+		return policy;
 	}
 
 	/**
@@ -332,20 +439,20 @@ export class TrustedTypeFactory extends EventTarget {
 		 * @Todo handle namespaced attributes
 		 */
 		if (typeof elementNS === 'string' && elementNs.length !== 0) {
-			return events.includes(attribute) ? TrustedScript.name : null;
+			return events.includes(attribute) ? 'TrustedScript' : null;
 		}
 
 		/**
 		 * This is an `on*` attribute
 		 */
 		if (events.includes(attribute)) {
-			return TrustedScript.name;
+			return 'TrustedScript';
 		}
 
 		switch(tagName) {
 			case 'script': {
 				if (attribute === 'src') {
-					return TrustedScriptURL.name;
+					return 'TrustedScriptURL';
 				} else {
 					return null;
 				}
@@ -353,9 +460,9 @@ export class TrustedTypeFactory extends EventTarget {
 
 			case 'iframe': {
 				if (attribute === 'srcdoc') {
-					return TrustedHTML.name;
+					return 'TrustedHTML';
 				} else if (attribute === 'src') {
-					return TrustedScriptURL.name;
+					return 'TrustedScriptURL';
 				} else {
 					return null;
 				}
@@ -432,7 +539,7 @@ export class TrustedTypeFactory extends EventTarget {
 	 * @return {TrustedTypePolicy} [description]
 	 */
 	get defaultPolicy() {
-		return getPolicy('default');
+		return this[symbols.defaultPolicy];
 	}
 
 	/**
@@ -452,13 +559,17 @@ export class TrustedTypeFactory extends EventTarget {
  * [trustedTypes description]
  * @type {TrustedTypeFactory}
  */
-export const trustedTypes = new TrustedTypeFactory(symbols.trustedKey);
+export const trustedTypes = new TrustedTypePolicyFactory(symbols.trustedKey);
 
 /**
  * [polyfill description]
  * @return {[type]}                       [description]
  */
 export function polyfill() {
+	if (! ('TrustedTypePolicyFactory' in globalThis)) {
+		globalThis.TrustedTypePolicyFactory = TrustedTypePolicyFactory;
+	}
+
 	if (! ('TrustedTypePolicy' in globalThis)) {
 		globalThis.TrustedTypePolicy = TrustedTypePolicy;
 	}